Crate srml_babe

Consensus extension module for BABE consensus.

Re-exports

pub use timestamp;

Structs

Authorities
Current epoch authorities.

CurrentSlot
Current slot number.

EpochIndex
Current epoch index.

EpochStartSlot
Slot at which the current epoch started. It is possible that no block was authored at the given slot and the epoch change was signalled later than this.

GenesisConfig

InherentDataProvider
Provides the slot duration inherent data for BABE.

Module

Randomness
The epoch randomness for the current epoch.

Enums

Call
The BABE SRML module

Constants

INHERENT_IDENTIFIER
The BABE inherent identifier.

PUBLIC_KEY_LENGTH
The length of the public key

RANDOMNESS_LENGTH
The length of the BABE randomness

VRF_OUTPUT_LENGTH
The length of the VRF output

Traits

BabeInherentData
Auxiliary trait to extract BABE inherent data.

Trait

Type Definitions

AuthorityId
A Babe authority identifier. Necessarily equivalent to the schnorrkel public key used in the main Babe module. If that ever changes, then this must, too.

BabeKey
A BABE public key

InherentType
The type of the BABE inherent.